requirements.txt
boto3
pytest
pylint
python-dotenv
psycopg2-binary
confluent-kafka
